+++ R/data.R
@@ -1,0 +1,9 @@
+#' Differential Expression
+#'
+#' A differential expression table
+"limma"
+
+#' Co-expression Modules
+#'
+#' A list of co-expression modules
+"wgcna"